Separated plot selection dropdown menu to new component for reuseability and to clean up code. Separated data to content/localPlotData.ts to clean up the components and pages a bit more. Fixed the top part of the plot page to not have multiple headers and look less messy.
