+++ next-app/src/components/SelectionTabComponent.tsx
@@ -1,10 +1,6 @@
 'use client';
 
 import { ReactElement } from 'react';
-import {
-    BODY_CLASSES,
-    H_1,
-} from '@/constants';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { IPaths } from '@/interfaces/types';
@@ -12,14 +8,12 @@ import { IPaths } from '@/interfaces/types';
 export default function SelectionTabComponent(prop: {
     paths: IPaths;
   }): ReactElement {
-    const pageTitle: string = "Population plots";
 
     const currentRoute = usePathname();
     
     return (
         <>
-            <div className={BODY_CLASSES}>
-                <div className={H_1}>{pageTitle}</div>
+            <div>
                 <div role="tablist" className="tabs tabs-lifted pb-4">
                     {Object.keys(prop.paths).map( key => (
                         <Link
