+++ next-app/src/content/localPlotData.ts
@@ -1,0 +1,802 @@
+  // Define superpopulations and their corresponding colors
+export const superPopulations: string[] = ["AFR", "EUR", "EAS", "SAS", "AMR"];
+
+export const superPopulationColorsDict = {
+    AFR: "#f25c5c",
+    AMR: "#dab862",
+    EAS: "#70c265",
+    EUR: "#5480f0",
+    SAS: "#999999",
+  };
+
+  // Define populations and their corresponding color
+export const subPopulations: string[] = [
+    "ACB",
+    "ASW",
+    "ESN",
+    "GWD",
+    "LWK",
+    "MSL",
+    "YRI",
+    "FIN",
+    "GBR",
+    "IBS",
+    "TSI",
+    "CDX",
+    "CHB",
+    "CHS",
+    "JPT",
+    "KHV",
+    "BEB",
+    "GIH",
+    "ITU",
+    "PJL",
+    "STU",
+    "CLM",
+    "MXL",
+    "PEL",
+    "PUR",
+  ];
+
+export const populationSubsets = [
+    {
+      "population": "ACB",
+      "superpopulation": "AFR"
+    },
+    {
+      "population": "ASW",
+      "superpopulation": "AFR"
+    },
+    {
+      "population": "BEB",
+      "superpopulation": "SAS"
+    },
+    {
+      "population": "PJL",
+      "superpopulation": "SAS"
+    },
+    {
+      "population": "STU",
+      "superpopulation": "SAS"
+    },
+    {
+      "population": "ITU",
+      "superpopulation": "SAS"
+    },
+    {
+      "population": "GIH",
+      "superpopulation": "SAS"
+    },
+    {
+      "population": "CDX",
+      "superpopulation": "EAS"
+    },
+    {
+      "population": "CHS",
+      "superpopulation": "EAS"
+    },
+    {
+      "population": "KHV",
+      "superpopulation": "EAS"
+    },
+    {
+      "population": "CHB",
+      "superpopulation": "EAS"
+    },
+    {
+      "population": "JPT",
+      "superpopulation": "EAS"
+    },
+    {
+      "population": "CLM",
+      "superpopulation": "AMR"
+    },
+    {
+      "population": "ESN",
+      "superpopulation": "AFR"
+    },
+    {
+      "population": "GWD",
+      "superpopulation": "AFR"
+    },
+    {
+      "population": "MSL",
+      "superpopulation": "AFR"
+    },
+    {
+      "population": "YRI",
+      "superpopulation": "AFR"
+    },
+    {
+      "population": "FIN",
+      "superpopulation": "EUR"
+    },
+    {
+      "population": "GBR",
+      "superpopulation": "EUR"
+    },
+    {
+      "population": "IBS",
+      "superpopulation": "EUR"
+    },
+    {
+      "population": "LWK",
+      "superpopulation": "AFR"
+    },
+    {
+      "population": "TSI",
+      "superpopulation": "EUR"
+    },
+    {
+      "population": "MXL",
+      "superpopulation": "AMR"
+    },
+    {
+      "population": "PEL",
+      "superpopulation": "AMR"
+    },
+    {
+      "population": "PUR",
+      "superpopulation": "AMR"
+    }
+  ]
+
+export const sampleAlleleDataGenomicPlot = {
+    'IGHV1-2*02_S4953': 
+    {
+      'superpopulation':
+      [
+        {
+          "frequency": 0.031073446327683617,
+          "n": 22,
+          "population": "AFR"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "EUR"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "EAS"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "SAS"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "AMR"
+        }
+      ],
+      'population':
+      [
+        {
+          "frequency": 0.0379746835443038,
+          "n": 3,
+          "population": "ACB"
+        },
+        {
+          "frequency": 0.03225806451612903,
+          "n": 2,
+          "population": "ASW"
+        },
+        {
+          "frequency": 0.035398230088495575,
+          "n": 4,
+          "population": "ESN"
+        },
+        {
+          "frequency": 0.008403361344537815,
+          "n": 1,
+          "population": "GWD"
+        },
+        {
+          "frequency": 0.01680672268907563,
+          "n": 2,
+          "population": "LWK"
+        },
+        {
+          "frequency": 0.02040816326530612,
+          "n": 2,
+          "population": "MSL"
+        },
+        {
+          "frequency": 0.06779661016949153,
+          "n": 8,
+          "population": "YRI"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "FIN"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "GBR"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "IBS"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "TSI"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "CDX"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "CHB"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "CHS"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "JPT"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "KHV"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "BEB"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "GIH"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "ITU"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "PJL"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "STU"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "CLM"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "MXL"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "PEL"
+        },
+        {
+          "frequency": 0.0,
+          "n": 0,
+          "population": "PUR"
+        }
+      ],
+      'SNPscore': "0.0",
+      'SNPsnips': [
+        "rs373342578(T:261,106452929)"
+      ],
+    },
+    'IGHV1-2*04':
+    {
+      'superpopulation':
+      [
+        {
+          "frequency": 0.211864406779661,
+          "n": 150,
+          "population": "AFR"
+        },
+        {
+          "frequency": 0.5942028985507246,
+          "n": 246,
+          "population": "EUR"
+        },
+        {
+          "frequency": 0.40925925925925927,
+          "n": 221,
+          "population": "EAS"
+        },
+        {
+          "frequency": 0.387037037037037,
+          "n": 209,
+          "population": "SAS"
+        },
+        {
+          "frequency": 0.4142857142857143,
+          "n": 116,
+          "population": "AMR"
+        }
+      ],
+      'population':
+      [
+        {
+          "frequency": 0.12658227848101267,
+          "n": 10,
+          "population": "ACB"
+        },
+        {
+          "frequency": 0.3387096774193548,
+          "n": 21,
+          "population": "ASW"
+        },
+        {
+          "frequency": 0.1592920353982301,
+          "n": 18,
+          "population": "ESN"
+        },
+        {
+          "frequency": 0.29411764705882354,
+          "n": 35,
+          "population": "GWD"
+        },
+        {
+          "frequency": 0.2689075630252101,
+          "n": 32,
+          "population": "LWK"
+        },
+        {
+          "frequency": 0.1836734693877551,
+          "n": 18,
+          "population": "MSL"
+        },
+        {
+          "frequency": 0.13559322033898305,
+          "n": 16,
+          "population": "YRI"
+        },
+        {
+          "frequency": 0.65,
+          "n": 65,
+          "population": "FIN"
+        },
+        {
+          "frequency": 0.7722772277227723,
+          "n": 78,
+          "population": "GBR"
+        },
+        {
+          "frequency": 0.40404040404040403,
+          "n": 40,
+          "population": "IBS"
+        },
+        {
+          "frequency": 0.5526315789473685,
+          "n": 63,
+          "population": "TSI"
+        },
+        {
+          "frequency": 0.44,
+          "n": 55,
+          "population": "CDX"
+        },
+        {
+          "frequency": 0.36363636363636365,
+          "n": 40,
+          "population": "CHB"
+        },
+        {
+          "frequency": 0.43,
+          "n": 43,
+          "population": "CHS"
+        },
+        {
+          "frequency": 0.4,
+          "n": 42,
+          "population": "JPT"
+        },
+        {
+          "frequency": 0.41,
+          "n": 41,
+          "population": "KHV"
+        },
+        {
+          "frequency": 0.4077669902912621,
+          "n": 42,
+          "population": "BEB"
+        },
+        {
+          "frequency": 0.5504587155963303,
+          "n": 60,
+          "population": "GIH"
+        },
+        {
+          "frequency": 0.4732142857142857,
+          "n": 53,
+          "population": "ITU"
+        },
+        {
+          "frequency": 0.28703703703703703,
+          "n": 31,
+          "population": "PJL"
+        },
+        {
+          "frequency": 0.21296296296296297,
+          "n": 23,
+          "population": "STU"
+        },
+        {
+          "frequency": 0.5142857142857142,
+          "n": 36,
+          "population": "CLM"
+        },
+        {
+          "frequency": 0.44285714285714284,
+          "n": 31,
+          "population": "MXL"
+        },
+        {
+          "frequency": 0.4142857142857143,
+          "n": 29,
+          "population": "PEL"
+        },
+        {
+          "frequency": 0.2857142857142857,
+          "n": 20,
+          "population": "PUR"
+        }
+      ],
+      'SNPscore': "0.0",
+      'SNPsnips': [
+        "rs112806369(A:98,106452766)"
+      ],
+    },
+    'IGHV1-2*06':
+    {
+      'superpopulation':
+      [
+        {
+          "frequency": 0.211864406779661,
+          "n": 150,
+          "population": "AFR"
+        },
+        {
+          "frequency": 0.1473429951690821,
+          "n": 61,
+          "population": "EUR"
+        },
+        {
+          "frequency": 0.07962962962962963,
+          "n": 43,
+          "population": "EAS"
+        },
+        {
+          "frequency": 0.21851851851851853,
+          "n": 118,
+          "population": "SAS"
+        },
+        {
+          "frequency": 0.16428571428571428,
+          "n": 46,
+          "population": "AMR"
+        }
+      ],
+      'population':
+      [
+        {
+          "frequency": 0.0759493670886076,
+          "n": 6,
+          "population": "ACB"
+        },
+        {
+          "frequency": 0.1774193548387097,
+          "n": 11,
+          "population": "ASW"
+        },
+        {
+          "frequency": 0.20353982300884957,
+          "n": 23,
+          "population": "ESN"
+        },
+        {
+          "frequency": 0.2689075630252101,
+          "n": 32,
+          "population": "GWD"
+        },
+        {
+          "frequency": 0.17647058823529413,
+          "n": 21,
+          "population": "LWK"
+        },
+        {
+          "frequency": 0.3163265306122449,
+          "n": 31,
+          "population": "MSL"
+        },
+        {
+          "frequency": 0.22033898305084745,
+          "n": 26,
+          "population": "YRI"
+        },
+        {
+          "frequency": 0.16,
+          "n": 16,
+          "population": "FIN"
+        },
+        {
+          "frequency": 0.1782178217821782,
+          "n": 18,
+          "population": "GBR"
+        },
+        {
+          "frequency": 0.16161616161616163,
+          "n": 16,
+          "population": "IBS"
+        },
+        {
+          "frequency": 0.09649122807017543,
+          "n": 11,
+          "population": "TSI"
+        },
+        {
+          "frequency": 0.048,
+          "n": 6,
+          "population": "CDX"
+        },
+        {
+          "frequency": 0.07272727272727272,
+          "n": 8,
+          "population": "CHB"
+        },
+        {
+          "frequency": 0.03,
+          "n": 3,
+          "population": "CHS"
+        },
+        {
+          "frequency": 0.18095238095238095,
+          "n": 19,
+          "population": "JPT"
+        },
+        {
+          "frequency": 0.07,
+          "n": 7,
+          "population": "KHV"
+        },
+        {
+          "frequency": 0.21359223300970873,
+          "n": 22,
+          "population": "BEB"
+        },
+        {
+          "frequency": 0.3302752293577982,
+          "n": 36,
+          "population": "GIH"
+        },
+        {
+          "frequency": 0.32142857142857145,
+          "n": 36,
+          "population": "ITU"
+        },
+        {
+          "frequency": 0.08333333333333333,
+          "n": 9,
+          "population": "PJL"
+        },
+        {
+          "frequency": 0.1388888888888889,
+          "n": 15,
+          "population": "STU"
+        },
+        {
+          "frequency": 0.2857142857142857,
+          "n": 20,
+          "population": "CLM"
+        },
+        {
+          "frequency": 0.11428571428571428,
+          "n": 8,
+          "population": "MXL"
+        },
+        {
+          "frequency": 0.12857142857142856,
+          "n": 9,
+          "population": "PEL"
+        },
+        {
+          "frequency": 0.12857142857142856,
+          "n": 9,
+          "population": "PUR"
+        }
+      ],
+      'SNPscore': "0.0",
+      'SNPsnips': [
+        "rs1065059(G:149,106452817)"
+      ],
+    }
+  }
+
+export const sampleAlleleDataAminoAcidPlot = {
+    'IGHV1-18*01_AA': 
+    {
+      'superpopulation':
+      [
+          {
+              "frequency": 0.8319209039548022,
+              "n": 589,
+              "population": "AFR"
+          },
+          {
+              "frequency": 0.9396135265700483,
+              "n": 389,
+              "population": "EUR"
+          },
+          {
+              "frequency": 0.9481481481481482,
+              "n": 512,
+              "population": "EAS"
+          },
+          {
+              "frequency": 0.7685185185185185,
+              "n": 415,
+              "population": "SAS"
+          },
+          {
+              "frequency": 0.8821428571428571,
+              "n": 247,
+              "population": "AMR"
+          }
+      ],
+      'population':
+      [
+          {
+              "frequency": 0.5822784810126582,
+              "n": 46,
+              "population": "ACB"
+          },
+          {
+              "frequency": 0.8064516129032258,
+              "n": 50,
+              "population": "ASW"
+          },
+          {
+              "frequency": 0.9292035398230089,
+              "n": 105,
+              "population": "ESN"
+          },
+          {
+              "frequency": 0.8739495798319328,
+              "n": 104,
+              "population": "GWD"
+          },
+          {
+              "frequency": 0.8487394957983193,
+              "n": 101,
+              "population": "LWK"
+          },
+          {
+              "frequency": 0.7244897959183674,
+              "n": 71,
+              "population": "MSL"
+          },
+          {
+              "frequency": 0.9491525423728814,
+              "n": 112,
+              "population": "YRI"
+          },
+          {
+              "frequency": 1.0,
+              "n": 100,
+              "population": "FIN"
+          },
+          {
+              "frequency": 0.9801980198019802,
+              "n": 99,
+              "population": "GBR"
+          },
+          {
+              "frequency": 0.8080808080808081,
+              "n": 80,
+              "population": "IBS"
+          },
+          {
+              "frequency": 0.9649122807017544,
+              "n": 110,
+              "population": "TSI"
+          },
+          {
+              "frequency": 0.984,
+              "n": 123,
+              "population": "CDX"
+          },
+          {
+              "frequency": 0.8454545454545455,
+              "n": 93,
+              "population": "CHB"
+          },
+          {
+              "frequency": 0.94,
+              "n": 94,
+              "population": "CHS"
+          },
+          {
+              "frequency": 0.9809523809523809,
+              "n": 103,
+              "population": "JPT"
+          },
+          {
+              "frequency": 0.99,
+              "n": 99,
+              "population": "KHV"
+          },
+          {
+              "frequency": 0.9805825242718447,
+              "n": 101,
+              "population": "BEB"
+          },
+          {
+              "frequency": 0.9908256880733946,
+              "n": 108,
+              "population": "GIH"
+          },
+          {
+              "frequency": 0.9464285714285714,
+              "n": 106,
+              "population": "ITU"
+          },
+          {
+              "frequency": 0.49074074074074076,
+              "n": 53,
+              "population": "PJL"
+          },
+          {
+              "frequency": 0.4351851851851852,
+              "n": 47,
+              "population": "STU"
+          },
+          {
+              "frequency": 1.0,
+              "n": 70,
+              "population": "CLM"
+          },
+          {
+              "frequency": 0.8285714285714286,
+              "n": 58,
+              "population": "MXL"
+          },
+          {
+              "frequency": 0.9857142857142858,
+              "n": 69,
+              "population": "PEL"
+          },
+          {
+              "frequency": 0.7142857142857143,
+              "n": 50,
+              "population": "PUR"
+          }
+      ],
+      'alleleListAA': [
+          "IGHV1-18*01",
+          "IGHV1-18*01_S0898",
+          "IGHV1-18*01_S1946",
+          "IGHV1-18*04"
+      ],
+    },
+  }
